requirements.txt
asgiref==3.5.2
Django==4.1.3
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-translated-fields==0.12.0
django-utils-six==2.0
django-widget-tweaks==1.4.12
jdatetime==4.1.0
Pillow==9.3.0
python-decouple==3.6
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
